% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpp1.R
\name{lpp1}
\alias{lpp1}
\title{Point pattern on single line segment}
\usage{
lpp1(x, domain, marks = NULL)
}
\arguments{
\item{x}{numeric vector of points on the line segment}

\item{domain}{numeric of length 2 with start and end of line segment}

\item{marks}{vector or \code{data.frame} with marks for the points given in \code{x}.}
}
\value{
object of class \code{lpp1} and \code{lpp}.
}
\description{
Point pattern on single line segment
}
\examples{
L <- c(0, 5)
x <- runif(10, L[1], L[2])
m <- data.frame(a = sample(letters, length(x)), b = seq_along(x))
X <- lpp1(x, domain = L, marks = m)
}
