% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat.1d-package.R
\docType{package}
\name{spatstat.1d-package}
\alias{spatstat.1d}
\alias{spatstat.1d-package}
\title{spatstat.1d: Spatial point pattern analysis in one dimension}
\description{
Implements some standard spatstat functionality to the simple case of point patterns in one dimension, that is, on a single segment of the real line.
}
\author{
\strong{Maintainer}: Ege Rubak \email{rubak@math.aau.dk} (0000-0002-6675-533X)

Authors:
\itemize{
  \item Ottmar Cronie \email{ottmar.cronie@umu.se} (0000-0002-6721-8608)
}

}
\keyword{internal}
